// src/utils/open-im-sdk-wasm/types/entity.d.ts
export declare type WSEvent<T = unknown> = {
    event: CbEvents;
    data: T;
    errCode: number;
    errMsg: string;
    operationID: string;
};
export declare type WsResponse<T = string> = {
    event: string;
    errCode: number;
    errMsg: string;
    data: T;
    operationID: string;
};
// ...
export declare type GroupApplicationItem = {
    createTime: number;
    creatorUserID: string;
    ex: string;
    gender: number;
    groupFaceURL: string;
    groupID: string;
    groupName: string;
    groupType: GroupType;
    handleResult: ApplicationHandleResult;
    handleUserID: string;
    handledMsg: string;
    handledTime: number;
    introduction: string;
    memberCount: number;
    nickname: string;
    notification: string;
    ownerUserID: string;
    reqMsg: string;
    reqTime: number;
    joinSource: GroupJoinSource;
    status: GroupStatus;
    userFaceURL: string;
    userID: string;
};
export declare type FriendApplicationItem = {
    createTime: number;
    ex: string;
    fromFaceURL: string;
    fromGender: number;
    fromNickname: string;
    fromUserID: string;
    handleMsg: string;
    handleResult: ApplicationHandleResult;
    handleTime: number;
    handlerUserID: string;
    reqMsg: string;
    toFaceURL: string;
    toGender: number;
    toNickname: string;
    toUserID: string;
};
export declare type FullUserItem = {
    blackInfo: BlackUserItem | null;
    friendInfo: FriendUserItem | null;
    publicInfo: PublicUserItem | null;
};
export declare type PublicUserItem = {
    nickname: string;
    userID: string;
    faceURL: string;
    ex: string;
};
export declare type SelfUserInfo = {
    createTime: number;
    ex: string;
    faceURL: string;
    nickname: string;
    userID: string;
    globalRecvMsgOpt: MessageReceiveOptType;
};
export declare type PartialUserInfo = {
    userID: string;
} & Partial<Omit<SelfUserInfo, 'userID'>>;
export declare type FriendUserItem = {
    addSource: number;
    createTime: number;
    ex: string;
    faceURL: string;
    userID: string;
    nickname: string;
    operatorUserID: string;
    ownerUserID: string;
    remark: string;
    attachedInfo: string;
};
export declare type SearchedFriendsInfo = FriendUserItem & {
    relationship: Relationship;
};
export declare type FriendshipInfo = {
    result: number;
    userID: string;
};
export declare type BlackUserItem = {
    addSource: number;
    userID: string;
    createTime: number;
    ex: string;
    faceURL: string;
    gender: number;
    nickname: string;
    operatorUserID: string;
    ownerUserID: string;
};
export declare type GroupItem = {
    groupID: string;
    groupName: string;
    notification: string;
    notificationUserID: string;
    notificationUpdateTime: number;
    introduction: string;
    faceURL: string;
    ownerUserID: string;
    createTime: number;
    memberCount: number;
    status: GroupStatus;
    creatorUserID: string;
    groupType: GroupType;
    needVerification: GroupVerificationType;
    ex: string;
    applyMemberFriend: AllowType;
    lookMemberInfo: AllowType;
};
export declare type GroupMemberItem = {
    groupID: string;
    userID: string;
    nickname: string;
    faceURL: string;
    roleLevel: GroupMemberRole;
    muteEndTime: number;
    joinTime: number;
    joinSource: GroupJoinSource;
    inviterUserID: string;
    operatorUserID: string;
    ex: string;
};
export declare type ConversationItem = {
    conversationID: string;
    conversationType: SessionType;
    userID: string;
    groupID: string;
    showName: string;
    faceURL: string;
    recvMsgOpt: MessageReceiveOptType;
    unreadCount: number;
    groupAtType: GroupAtType;
    latestMsg: string;
    latestMsgSendTime: number;
    draftText: string;
    draftTextTime: number;
    burnDuration: number;
    msgDestructTime: number;
    isPinned: boolean;
    isNotInGroup: boolean;
    isPrivateChat: boolean;
    isMsgDestruct: boolean;
    attachedInfo: string;
    ex: string;
};
// ...
export declare type RevokedInfo = {
    revokerID: string;
    revokerRole: number;
    clientMsgID: string;
    revokerNickname: string;
    revokeTime: number;
    sourceMessageSendTime: number;
    sourceMessageSendID: string;
    sourceMessageSenderNickname: string;
    sessionType: number;
    seq: number;
    ex: string;
};